Fix msgPositions map key to Long

+++ app/src/main/java/com/feder/compose/ui/screen/ChatScreen.kt
@@ -368,7 +368,7 @@ fun ChatScreen(chatName: String, chatUsername: String, myUsername: String, token
         }
     }
 
-    val msgPositions = remember { mutableMapOf<Int, androidx.compose.ui.geometry.Offset>() }
+    val msgPositions = remember { mutableMapOf<Long, androidx.compose.ui.geometry.Offset>() }
 
     val dateInHeader = remember { mutableStateOf("") }
